My first site
There are fonts on the project that I did not use, I upload it to github a month later, that is, I learned new skills but decided not to fix anything since this is my first site. I left everything as it is, let my progress be visible with each site.